+++ app/Http/Controllers/ClientController.php
@@ -293,7 +293,7 @@ class ClientController extends Controller
         ->join('forms', 'forms.form_code', '=', 'form_id')
         ->join('merchants', 'merchants.id', '=', 'forms.merchant_id')
         // ->join('attachments','attachments.submission_code', '=', 'submitted_forms.submission_code')
-        ->select('submitted_forms.*','merchants.merchant_name AS merchant_name', 'merchants.nickname',
+        ->select('submitted_forms.*','merchants.merchant_name AS merchant_name', 'merchants.nickname', 'merchants.id AS merchant_id',
         'users.name', 'users.email', 'forms.name AS form_name', 'forms.form_fields', 'logo')
         ->where('submitted_forms.client_id', $id)
         ->where('submitted_forms.status', '!=', 4)
@@ -310,6 +310,7 @@ class ClientController extends Controller
             $submittedformdata['form_name'] = Crypt::decryptString($items->form_name);
             $submittedformdata['form_fields'] = json_decode(Crypt::decryptString($items->form_fields));
             $submittedformdata['merchant_name'] = Crypt::decryptString($items->merchant_name);
+            $submittedformdata['merchant_id'] = $items->merchant_id;
             $submittedformdata['nickname'] = $items->nickname;
             $submittedformdata['client_name'] = $items->name;
             $submittedformdata['email'] = $items->email;