+++ resources/forms-ui/src/app/pages/dashboard/git-admin/edit-access-code-page/edit-access-code-page.component.ts
@@ -153,7 +153,7 @@ export class EditAccessCodePageComponent implements OnInit {
   getCompanyBranches() {
     const id = this.getSelectedMerchantIdentifier();
     console.log('selected company_id: ' + id);
-    this.branchService.getBranch(id.toString()).then(
+    this.branchService.getCompanyBranches(id.toString()).then(
       res => {
         const branches = res as any;
         console.log('res_length: ' + branches.length);